+++ Kernel/Net/LoopbackAdapter.cpp
@@ -44,7 +44,7 @@ LoopbackAdapter::~LoopbackAdapter()
 {
 }
 
-void LoopbackAdapter::send_raw(const u8* data, int size)
+void LoopbackAdapter::send_raw(const u8* data, size_t size)
 {
     dbgprintf("LoopbackAdapter: Sending %d byte(s) to myself.\n", size);
     did_receive(data, size);
